Update to address reviewable comments
