+++ pom.xml
@@ -23,8 +23,6 @@
     <scala.version.prefix>2.10</scala.version.prefix>
     <hadoop.version>2.3.0</hadoop.version>
     <spark.version>1.3.1</spark.version>
-    <scoverage-plugin.version>0.99.11</scoverage-plugin.version>
-    <maven.plugin.scoverage.version>0.99.10</maven.plugin.scoverage.version>
     <maven.release.plugin.version>2.4.1</maven.release.plugin.version>
   </properties>
 
@@ -205,11 +203,6 @@
         </executions>
       </plugin>
       <plugin>
-        <groupId>org.scoverage</groupId>
-        <artifactId>maven-scoverage-plugin</artifactId>
-        <version>${maven.plugin.scoverage.version}</version>
-      </plugin>
-      <plugin>
         <groupId>org.apache.maven.plugins</groupId>
         <artifactId>maven-release-plugin</artifactId>
         <version>${maven.release.plugin.version}</version>
@@ -358,11 +351,6 @@
       <version>0.6.0</version>
     </dependency>
     <dependency>
-      <groupId>org.scoverage</groupId>
-      <artifactId>scalac-scoverage-plugin_${scala.version.prefix}</artifactId>
-      <version>${scoverage-plugin.version}</version>
-    </dependency>
-    <dependency>
       <groupId>com.assembla.scala-incubator</groupId>
       <artifactId>graph-core_2.10</artifactId>
       <version>1.9.1</version>
@@ -396,21 +384,11 @@
                 <configuration>
                   <args>
                     <arg>-g:vars</arg>
-                    <!-- <arg>-P:scoverage:dataDir:${project.build.directory}</arg> -->
                   </args>
                 </configuration>
               </execution>
             </executions>
             <configuration>
-            <!--
-              <compilerPlugins>
-                <compilerPlugin>
-                  <groupId>org.scoverage</groupId>
-                  <artifactId>scalac-scoverage-plugin_${scala.version.prefix}</artifactId>
-                  <version>${scoverage-plugin.version}</version>
-                </compilerPlugin>
-              </compilerPlugins>
-            -->
             </configuration>
           </plugin>
         </plugins>
